Move GC tetscases to GC module
